part way through refactoring `documents`
